FastAPI-Docker/requirements.txt
absl-py==1.2.0
anyio==3.6.1
astunparse==1.6.3
asyncpg==0.26.0
bcrypt==3.2.2
blinker==1.4
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
databases==0.6.1
DateTime==4.7
fastapi==0.85.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
flatbuffers==22.9.24
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.49.1
h11==0.14.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
protobuf==3.19.6
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.39
starlette==0.20.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.12
uuid==1.30
uvicorn==0.18.3
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
zope.interface==5.5.0
